producer/requirements.txt
requests==2.32.3
kafka-python==2.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
prometheus-client==0.20.0

